@@ -1,16 +1,15 @@
 class MidiHeader {
-  final int framesPerSecond;
+  final int? framesPerSecond;
   final int ticksPerBeat;
-  final int ticksPerFrame;
+  final int? ticksPerFrame;
   final int numTracks;
-  final int format;
-  final int timeDivision;
+  final int? format;
+  final int? timeDivision;
 
   MidiHeader(
-      {this.framesPerSecond,
-      this.ticksPerBeat,
-      this.format,
-      this.numTracks,
-      this.ticksPerFrame,
+      { this.framesPerSecond,
+      required this.ticksPerBeat, this.format,
+      required this.numTracks,
+        this.ticksPerFrame,
       this.timeDivision});
 }
